requirements.txt
asgiref==3.2.10
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.1.1
djangorestframework==3.11.1
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
